robinhoodAPI updates to log execution price
Several people in the group where I learned about this script were talking about functionality to update their records with order execution fills, so I was wondering if this was something that you'd want to include in the main script. 

This is a fairly straightforward implementation, it just adds save_execution_price() as a csv file. Ideally this functionality would be added to the other APIs as well. Thanks!