Updated discord bot code
Removed any references I could find restricting the discord bot to a specific channel.

I swapped it to utilize slash (/) commands and operate under a user installed bot rather than a guild installed bot. Discord has this as an official option when creating a bot and I felt having to have an entire server dedicated for a single bot was unnecessary so now it can be used anywhere at all.

All messages are sent with ephemeral set to true if not in DMs with the bot so the bot commands can be executed in literally any server or in a DM with the bot itself and nobody else can see it.

If commands are executed in a server, responses for holdings and transactions are DMed to you by the bot instead of sent in the server, even though ephemeral would be set to true so they would be hidden I felt it'd be better to have it in DMs either way.

Holdings command is tested and working, Transactions command should function just as normal but have not fully tested since the market is closed. I did try it and it returned errors properly about what ticker I chose, the quantity, and the broker, but it would be worth testing again while the market is open before fully releasing.

Dry is set to a default of True for transaction commands unless overrode during executing the command.